Refactor Questions into Data Directory
Moves the questions data out of App.jsx and into a dedicated data directory.

* Refactor all questions and move each to their own files and constants
* Move modules, modules info, and recommendations to their own files and constants
* Import refactored files into App.jsx
* Update Site Title
* Fix Production Base URL in Pages deploy
* Update Staging URL assignment in GH Action
