+++ .github/workflows/pages.yml
@@ -5,9 +5,6 @@ on:
   push:
     branches: ['main']
 
-env:
-  BASE: '/'
-
 
 permissions:
   contents: read
@@ -32,16 +29,28 @@ jobs:
         with:
           node-version: lts/*
           cache: 'npm'
+
+      - name: Set Build Variables
+        run: |
+          REPO_NAME="${GITHUB_REPOSITORY##*/}"
+          echo "REPO_NAME=$REPO_NAME" >> $GITHUB_ENV
+          echo "BASE=/$REPO_NAME/" >> "$GITHUB_ENV"
+      
+
       - name: Install dependencies
         run: npm install
+      
       - name: Build
         run: npm run build
+      
       - name: Setup Pages
         uses: actions/configure-pages@v5
+      
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
           path: './dist'
+      
       - name: Deploy to GitHub Pages
         id: deployment
         uses: actions/deploy-pages@v4
