+++ .github/workflows/preview.yml
@@ -10,8 +10,6 @@ on:
 env:
   PR_NUMBER: ${{ github.event.number }}
 
-
-
 # Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
 permissions:
     contents: read
@@ -26,19 +24,23 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
 
-      - run: echo "PREVIEW_PATH=pr-preview/pr-${{ github.event.number }}" >> "$GITHUB_ENV"
-
       - uses: actions/setup-node@v4
       - name: Install Node Dependencies
         run: npm install
 
+      - name: Set Build Variables
+        run: |
+          REPO_NAME="${GITHUB_REPOSITORY##*/}"
+          echo "REPO_NAME=$REPO_NAME" >> $GITHUB_ENV
+          echo "BASE=/$REPO_NAME/pr-preview/pr-${{ github.event.number }}" >> "$GITHUB_ENV"
+
       - name: Install and Build
         if: github.event.action != 'closed' # Skip the build if the PR has been closed
-        env:
-          BASE: '/pr-preview/pr-${{ github.event.number }}'
+
         run: npm run build
 
       - name: Deploy preview
         uses: rossjrw/pr-preview-action@v1
         with:
           source-dir: ./dist/
+
