+++ src/data/modules.jsx
@@ -1,0 +1,8 @@
+const AI_MODULES = {
+  MAPPING: 'ai-mapping',
+  REGULATION: 'ai-regulation',
+  RESPONSIBLE_AI: 'responsible-ai',
+  RISK: 'ai-risk'
+};
+
+export default AI_MODULES;
